requirements.txt
python-fasthtml
datetime
pandas
sqlite_minutils
pyarrow
requests
pytz